% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commodities.R
\name{add_target_use}
\alias{add_target_use}
\title{Find nearest LLIN target use to match to NPC outputs}
\usage{
add_target_use(x)
}
\arguments{
\item{x}{Model output}
}
\description{
Find nearest LLIN target use to match to NPC outputs
}
