% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commodities.R
\name{adjust_net_efficiency}
\alias{adjust_net_efficiency}
\title{After 2023, net distribution switches to the most efficient observed country for all countries}
\usage{
adjust_net_efficiency(x)
}
\arguments{
\item{x}{Model output}
}
\description{
After 2023, net distribution switches to the most efficient observed country for all countries
}
