% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commodities.R
\name{annual_net_distibuted}
\alias{annual_net_distibuted}
\title{Annual nets distributed}
\usage{
annual_net_distibuted(eq_npc)
}
\arguments{
\item{eq_npc}{Equilibrium nets per capita}
}
\description{
The annual number of nets distributed to maintain the input equilibrium nets per capita.
This is for the specific "smoothed" case where we assume that 1/3 of the population at risk
is distributed nets every 3 years (this is equivalent to a smoothed 3 yearly whole population distribution).
}
