% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commodities.R
\name{annual_net_distibuted_gts}
\alias{annual_net_distibuted_gts}
\title{Annual nets distributed}
\usage{
annual_net_distibuted_gts(eq_npc)
}
\arguments{
\item{eq_npc}{Equilibrium nets per capita}
}
\description{
The annual number of nets distributed to maintain the input equilibrium nets per capita.
This is for the specific "smoothed" case where we assume that 1/3 of the population at risk
is distributed nets every 3 years (this is equivalent to a smoothed 3 yearly whole population distribution).
This version is a simplified approximation with 3 values for proportion of nets remaining (1, 2, 3 years) scraped from
\href{https://elifesciences.org/articles/09672#app1}{Bhatt \emph{et al}, 2015}. We use this approach for consistency with GTS.
For a full approach with more updated fits see \code{\link{annual_net_distibuted}}. This newer approach is slightly more
optimistic than this approach (although overall uncertainty in these projections is large).
}
