% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{any_vc_coverage}
\alias{any_vc_coverage}
\title{Add coverage of either net or IRS. These are modelled as -ve correlated so
the coverage is their sum (to max of 100\%)}
\usage{
any_vc_coverage(x)
}
\arguments{
\item{x}{Model output}
}
\description{
Add coverage of either net or IRS. These are modelled as -ve correlated so
the coverage is their sum (to max of 100\%)
}
