% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commodities.R
\name{approx_target_use}
\alias{approx_target_use}
\title{Linear interpolation for eq_npc use}
\usage{
approx_target_use(iso, eq_npc)
}
\arguments{
\item{iso}{ISO code}

\item{eq_npc}{Equilibrium NPC}
}
\description{
Linear interpolation for eq_npc use
}
