% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_inputs.R
\name{back_adjust_iccm}
\alias{back_adjust_iccm}
\title{Split out iccm coverage from treatment coverage}
\usage{
back_adjust_iccm(x)
}
\arguments{
\item{x}{Interventions data.frame}
}
\description{
Split out iccm coverage from treatment coverage
}
