% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{case_weighting}
\alias{case_weighting}
\title{Case weighting.}
\format{
An object of class \code{data.frame} with 64 rows and 2 columns.
}
\usage{
case_weighting
}
\description{
Ratio of deaths:cases from GP to use to downscale cases to weight them
equally to deaths in optimisation.
}
\keyword{datasets}
