% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commodities.R
\name{commodities_and_services}
\alias{commodities_and_services}
\title{Commodity and service estimates}
\usage{
commodities_and_services(x)
}
\arguments{
\item{x}{Model output}
}
\description{
We use the net model (see net-specific functions) for the number of LLINs.
We approximate microscopy->act transition using the proportion acts.
We assume, that for areas with significant ongoing transmission (prev > 5\%), 37\% of NMFs are tested with RDTs (assumming proprtion_act = 1) - this is informed from the DHS.
We further assume, that of those NMFs tested with RDTS and prevalence-dependent proportion return a +ve rdt and are treated.
}
