% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_inputs.R
\name{create_coverage}
\alias{create_coverage}
\title{Modify GP intervention data.frame for coverage options}
\usage{
create_coverage(tx, llin, irs, smc, rtss, iccm, ipti, gp_interventions)
}
\arguments{
\item{tx}{Binary indicator for treatment}

\item{llin}{Binary indicator for nets}

\item{irs}{Binary indicator for irs}

\item{smc}{Binary indicator for smc}

\item{rtss}{Binary indicator for rtss}

\item{iccm}{Binary indicator for iccm}

\item{ipti}{Binary indicator for ipti}

\item{gp_interventions}{intervention data.frame from the GP}
}
\description{
Modify GP intervention data.frame for coverage options
}
