% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_inputs.R
\name{create_intervention_option_matrix}
\alias{create_intervention_option_matrix}
\title{Create matrix of coverage options}
\usage{
create_intervention_option_matrix(ref_year)
}
\arguments{
\item{ref_year}{Interventions for the GP year 2026}
}
\description{
Binary indcator if intervention is to be implemented (at GP coverage) or not
}
