% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_inputs.R
\name{create_names}
\alias{create_names}
\title{Create the replenishment naming scheme indicating which interventions are on and off during the replenishment period}
\usage{
create_names(intervention_binary_options, names)
}
\arguments{
\item{intervention_binary_options}{Binary indicator if intervention is on or off}

\item{names}{Intervention names}
}
\description{
Create the replenishment naming scheme indicating which interventions are on and off during the replenishment period
}
