% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimisation.R
\name{create_optim_matrices}
\alias{create_optim_matrices}
\title{Create optimisation matrices}
\usage{
create_optim_matrices(x, nopt, nsites, maxy, maxcost)
}
\arguments{
\item{x}{Optimisation data filtered for dominant solutions}

\item{nopt}{Maximum number of options per site}

\item{nsites}{Number of sites}

\item{maxy}{Maximum y over all sites}

\item{maxcost}{Maximum cost over all sites}
}
\description{
Creates two i (solution) x j (site) matrices one containing costs one y.
Where the number of solutions between sites differ, "empty" solutions are
assigned maximum cost and maximum y (an will therefore never be selected)
}
