% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimisation.R
\name{create_optimisation_data}
\alias{create_optimisation_data}
\title{Create dataset for optimisation}
\usage{
create_optimisation_data(processed_output)
}
\arguments{
\item{processed_output}{Processed model output}
}
\value{
Data for use in optimisation
}
\description{
Create dataset for optimisation
}
