% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi.R
\name{dalys_cast_forward}
\alias{dalys_cast_forward}
\title{Dalys case forward}
\usage{
dalys_cast_forward(x, lifespan, ...)
}
\arguments{
\item{x}{Model output}

\item{lifespan}{Average life expectancy}

\item{...}{Grouping arguments}
}
\description{
Where, for example, a death of newborn in year 2000 with a life expectancy of 63 will add 1 YLL for each year from 2000:2062
}
