% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_post.R
\name{elim_threshold}
\alias{elim_threshold}
\title{Elimination threshold}
\usage{
elim_threshold(x, threshold = 0)
}
\arguments{
\item{x}{output}

\item{threshold}{Incidence threshold below which outputs are fixed}
}
\description{
Elimination threshold
}
