% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_inputs.R
\name{extend}
\alias{extend}
\title{Extend any inputs that do not span the full modelled time horizon}
\usage{
extend(x)
}
\arguments{
\item{x}{Input data.frame}
}
\description{
Extend any inputs that do not span the full modelled time horizon
}
