% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_inputs.R
\name{fixed_post}
\alias{fixed_post}
\title{Continue coverage in post replenishment period}
\usage{
fixed_post(interventions)
}
\arguments{
\item{interventions}{An interventions dataframe}
}
\description{
Continue coverage in post replenishment period
}
