% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_inputs.R
\name{forward_adjust_iccm}
\alias{forward_adjust_iccm}
\title{Combine iccm coverage with treatment coverage}
\usage{
forward_adjust_iccm(x)
}
\arguments{
\item{x}{Interventions data.frame}
}
\description{
Combine iccm coverage with treatment coverage
}
