% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf-package.R
\docType{package}
\name{gf-package}
\alias{gf}
\alias{gf-package}
\title{gf: What the Package Does (One Line, Title Case)}
\description{
For more information visit <https://2degrees-investing.org/>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://maurolepore.github.io/gf}
}

}
\author{
\strong{Maintainer}: Mauro Lepore \email{maurolepore@gmail.com} (\href{https://orcid.org/0000-0002-1986-7988}{ORCID}) [contractor]

Other contributors:
\itemize{
  \item 2 Degrees Investing Initiative \email{contact@2degrees-investing.org} [copyright holder, funder]
}

}
\keyword{internal}
