% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf_output.R
\name{gf_aggregate}
\alias{gf_aggregate}
\title{Aggregate output}
\usage{
gf_aggregate(x, ...)
}
\arguments{
\item{x}{Optimised output}

\item{...}{Grouping columns}
}
\description{
Aggregation can be over sites and/or age (not time).
}
