% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gf_countries}
\alias{gf_countries}
\title{List of countries to model for the GF}
\format{
An object of class \code{data.frame} with 65 rows and 2 columns.
}
\usage{
gf_countries
}
\description{
List of countries to model for the GF
}
\keyword{datasets}
