% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gf_public}
\alias{gf_public}
\title{Proportion of treatment public sector}
\format{
An object of class \code{data.frame} with 66 rows and 2 columns.
}
\usage{
gf_public
}
\description{
Proportion of treatment public sector
}
\keyword{datasets}
