% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commodities.R
\name{half_life}
\alias{half_life}
\title{Half life net loss function}
\usage{
half_life(k, l)
}
\arguments{
\item{k}{Rate (fixed from paper)}

\item{l}{Time at which all nets = 0 (l is estimated so that \code{\link{half_life}}  = 1.64 years (median retention span across all countries))}
}
\description{
Half life net loss function
}
