% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hh}
\alias{hh}
\title{Average HH size}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 80 rows and 2 columns.
}
\usage{
hh
}
\description{
Average HH size
}
\keyword{datasets}
