% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_inputs.R
\name{interventions_rename}
\alias{interventions_rename}
\title{Rename interventions}
\usage{
interventions_rename(x)
}
\arguments{
\item{x}{GP interventions data.frame}
}
\description{
Rename interventions
}
