% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi.R
\name{life_years}
\alias{life_years}
\title{Add Life years lived}
\usage{
life_years(x)
}
\arguments{
\item{x}{Model output}
}
\description{
Years of life lived in year Y for age group A (assuming deaths occur on average half way through the year)
}
