% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_data.R
\name{link_data}
\alias{link_data}
\title{Link simulation output with input data}
\usage{
link_data(x, coverage_input)
}
\arguments{
\item{x}{Model output}

\item{coverage_input}{Coverage inputs for run}
}
\description{
Link simulation output with input data
}
