% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi.R
\name{mortality_rate}
\alias{mortality_rate}
\title{Add mortality rate (dependent on severe case incidence and treatment coverage)}
\usage{
mortality_rate(x, scaler = 0.215, treatment_scaler = 0.5)
}
\arguments{
\item{x}{Model output}

\item{scaler}{Severe case to death scaler}

\item{treatment_scaler}{Treatment modifier}
}
\description{
This follows methodology in \href{https://pubmed.ncbi.nlm.nih.gov/26809816/}{Griffin \emph{et al}, 2016}.
}
