% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimisation.R
\name{multi_optimisation}
\alias{multi_optimisation}
\title{Run multiple optimisations}
\usage{
multi_optimisation(x, budget_prop)
}
\arguments{
\item{x}{Optimisation data}

\item{budget_prop}{Vector of proportions: these are the budgets as a proportion of the GP budget}
}
\value{
Optimisated output
}
\description{
Run multiple optimisations
}
