% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commodities.R
\name{net_loss}
\alias{net_loss}
\title{Net loss function}
\usage{
net_loss(t, k = 18, l = 8.5)
}
\arguments{
\item{t}{Time (years)}

\item{k}{Rate (fixed from paper)}

\item{l}{Time at which all nets = 0 (l is estimated so that \code{\link{half_life}}  = 1.64 years (median retention span across all countries))}
}
\value{
Proportion of nets retained
}
\description{
Adapted from:
\enumerate{
\item \href{https://elifesciences.org/articles/09672#app1}{Bhatt \emph{et al}, 2015}
\item \href{https://www.researchsquare.com/article/rs-199628/v1}{Bertozzi-Villa \emph{et al}, 2021 (Under review)}
}
}
