% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi.R
\name{outcome_uncertainty}
\alias{outcome_uncertainty}
\title{Add case and death uncertainty}
\usage{
outcome_uncertainty(x, cases_cv = 0.227, deaths_cv = 0.265)
}
\arguments{
\item{x}{Model output}

\item{cases_cv}{Case uncertainty SD scaler}

\item{deaths_cv}{Death uncertainty SD scaler}
}
\description{
Add case and death uncertainty
}
