% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_epi.R
\name{population_indicators}
\alias{population_indicators}
\title{Add some population-level indicators}
\usage{
population_indicators(x)
}
\arguments{
\item{x}{Model output}
}
\description{
These indicators aggregate prevalence and API summaries across age groups to inform
some strategy thresholds in costing. These are not summaries at the country level as
this would require re-running for each iteration of the optimisation, therefore they
are all pre-computed at the site level
}
