% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population_projections}
\alias{population_projections}
\title{Population projections}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 219024 rows and 11 columns.
}
\usage{
population_projections
}
\description{
Population and population at risk projections.
\href{https://ghsl.jrc.ec.europa.eu/index.php}{European Commission. GHSL - Global Human Settlement Layer}
Lloyd \emph{et al}. Data Descriptor: High resolution global gridded data for use in population studies. Nat Sci Data. 2017
\href{https://population.un.org/wpp/}{United Nations. World Population Prospects. 2019}
\href{https://population.un.org/wup/ }{United Nations. World Urbanization Prospects. 2018.}
}
\keyword{datasets}
