% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_post.R
\name{post_process}
\alias{post_process}
\title{Post processing wrapper}
\usage{
post_process(model_output, coverage_input)
}
\arguments{
\item{model_output}{Model output}

\item{coverage_input}{Coverage input}
}
\description{
Links to data, estimates epi, commodity and cost outputs
}
