% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ppf}
\alias{ppf}
\title{Proportion pf}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 208794 rows and 8 columns.
}
\usage{
ppf
}
\description{
Proportion of cases that are \emph{plasmodium falciparum}. Sourced from the WHO GMP. Proportion is assumed constant in the future.
}
\keyword{datasets}
