% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_raw.R
\name{process_raw}
\alias{process_raw}
\title{Raw output pre-processing wrapper}
\usage{
process_raw(raw_output, ...)
}
\arguments{
\item{raw_output}{Raw model output}

\item{...}{Additional columns to select (for example run names)}
}
\value{
Pre-processed model output
}
\description{
Raw output pre-processing wrapper
}
