% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_post.R
\name{prune}
\alias{prune}
\title{Keep only required columns - helps for memory issues with large countries}
\usage{
prune(x)
}
\arguments{
\item{x}{model_output Model output}
}
\description{
Keep only required columns - helps for memory issues with large countries
}
