% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_report.R
\name{render_report}
\alias{render_report}
\title{Create country summary report}
\usage{
render_report(
  gf_optimised,
  gf_gp,
  gts_gp,
  global_plan_budget,
  gf_fixed,
  comp,
  output_file,
  budget_levels = paste(seq(0.05, 1, 0.05))
)
}
\arguments{
\item{gf_optimised}{Optimised data for a country}

\item{gf_gp}{Country level annual summary from the GF runs of global plan}

\item{gts_gp}{Country level annual summary from the GTS global plan}

\item{global_plan_budget}{Replenishment period budget from the global plan}

\item{gf_fixed}{Fixed scenario output}

\item{comp}{Comparison with pre-inputs}

\item{output_file}{Report output address}

\item{budget_levels}{Budget increments used}
}
\description{
Create country summary report
}
