% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_inputs.R
\name{replenishment_options}
\alias{replenishment_options}
\title{Make replenishment options}
\usage{
replenishment_options(gp_input_single)
}
\arguments{
\item{gp_input_single}{A input row (site) from the GP}
}
\value{
data.frame of all options for the site
}
\description{
Make replenishment options
}
