% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimisation.R
\name{single_optimisation}
\alias{single_optimisation}
\title{Perform optimisation for a single budget line}
\usage{
single_optimisation(x, budget)
}
\arguments{
\item{x}{Optimisation data}

\item{budget}{Budget}
}
\value{
Optimisaed output
}
\description{
Perform optimisation for a single budget line
}
