% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{treatment_unit_costs}
\alias{treatment_unit_costs}
\title{Treatment unit costs}
\format{
An object of class \code{data.frame} with 3880 rows and 10 columns.
}
\usage{
treatment_unit_costs
}
\description{
A dataset containing the age- and country- specific (treatment) unit costs.
These were sources from a combination of costing from the original GTS
\href{https://gh.bmj.com/content/2/2/e000176}{Patouillard \emph{et al}, 2017} and an updated costing review (Ehler \emph{et al}, in prep).
and the WHOCHOICE data base \href{https://www.who.int/choice/cost-effectiveness/en/}{WHO CHOICE}.
}
\keyword{datasets}
