% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{unit_costs}
\alias{unit_costs}
\title{Unit costs}
\format{
An object of class \code{data.frame} with 1 rows and 14 columns.
}
\usage{
unit_costs
}
\description{
A dataset containing the non-country specific (non-treatment) unit costs.
These were sources from a combination of costing from the original GTS
\href{https://gh.bmj.com/content/2/2/e000176}{Patouillard \emph{et al}, 2017} and an updated costing review (Ehler \emph{et al}, in prep).
}
\keyword{datasets}
