% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{who_burden}
\alias{who_burden}
\title{WHO burden estimates (callibrated to)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2071 rows and 3 columns.
}
\usage{
who_burden
}
\description{
WHO burden estimates (callibrated to)
}
\keyword{datasets}
