% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi.R
\name{yll_cast_forward}
\alias{yll_cast_forward}
\title{Sum of deaths in last "lifetime left" years}
\usage{
yll_cast_forward(row, lifeleft, cumulative_deaths)
}
\arguments{
\item{row}{Row index}

\item{lifeleft}{Years of life left}

\item{cumulative_deaths}{Cumulative deaths}
}
\description{
Sum of deaths in last "lifetime left" years
}
