% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commodities.R
\name{adjust_net_efficiency}
\alias{adjust_net_efficiency}
\title{Find nearest LLIN target use to match to NPC outputs}
\usage{
adjust_net_efficiency(x)
}
\arguments{
\item{x}{Model output}
}
\description{
Find nearest LLIN target use to match to NPC outputs
}
