% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_inputs.R
\name{coverage_options}
\alias{coverage_options}
\title{Create set of binary coverage options}
\usage{
coverage_options(reference_coverage)
}
\arguments{
\item{reference_coverage}{GP coverage in 2026}
}
\description{
Return 0 if no coverage in GP reference year or 0, 1 if there was,
}
