% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_report.R
\name{render_report}
\alias{render_report}
\title{Create country summary report}
\usage{
render_report(gf_optimised, output_file)
}
\arguments{
\item{gf_optimised}{Optimised data for a country}

\item{output_file}{Report output address}
}
\description{
Create country summary report
}
